requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap4==3.0.1
django-countries==7.2.1
django-crispy-forms==1.12.0
django-datetimepicker==3.14
django-heroku==0.3.1
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
gunicorn==20.0.4
httplib2==0.18.1
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.3.0

